app:: trying to refactor with APProviderGL class (0)
