+++ app/src/main/java/good/damn/wrapper/providers/APIProviderGLRegister.kt
@@ -1,0 +1,9 @@
+package good.damn.wrapper.providers
+
+import good.damn.wrapper.models.APMProviderGL
+
+interface APIProviderGLRegister {
+    fun registerGlProvider(
+        provider: APMProviderGL
+    )
+}
