+++ app/src/main/java/good/damn/wrapper/providers/APProviderGL.kt
@@ -1,0 +1,13 @@
+package good.damn.wrapper.providers
+
+import good.damn.common.camera.COICameraFree
+import good.damn.engine2.models.MGMInformatorShader
+import good.damn.engine2.models.MGMManagers
+import good.damn.engine2.models.MGMParameters
+import good.damn.engine2.opengl.MGMGeometry
+import good.damn.engine2.opengl.pools.MGMPools
+import good.damn.wrapper.models.APMProviderGL
+
+open class APProviderGL {
+    lateinit var glProvider: APMProviderGL
+}
