+++ app/src/main/java/good/damn/wrapper/renderer/APRendererHandler.kt
@@ -70,9 +70,8 @@ import good.damn.wrapper.files.APFile
 import good.damn.wrapper.hud.APHud
 import java.util.concurrent.ConcurrentLinkedQueue
 
-class APRendererHandler(
-    requesterUserContent: APIRequestUserContent
-): GLSurfaceView.Renderer {
+class APRendererHandler
+: GLSurfaceView.Renderer {
 
     companion object {
         private const val TAG = "MGRendererLevelEditor"
@@ -80,251 +79,12 @@ class APRendererHandler(
 
     private val mHandlerGlExecutor = COHandlerGlExecutor()
 
-    private val mBuffer = ByteArray(
-        1024 * 50
-    )
-
     private val mHandlerGl = COHandlerGl(
         mHandlerGlExecutor.queue,
         mHandlerGlExecutor.queueCycle,
     )
 
-    private val mInformatorShader = MGMInformatorShader(
-        MGShaderSource(
-            "opaque",
-            mBuffer
-        ),
-        cacheGeometryPass = MGShaderCache(
-            SparseArray(50),
-            mHandlerGl,
-            GLShaderCreatorGeomPassModel()
-        ),
-        cacheGeometryPassInstanced = MGShaderCache(
-            SparseArray(50),
-            mHandlerGl,
-            GLShaderCreatorGeomPassInstanced()
-        ),
-        wireframe = GLShaderGeometryPassModel(
-            GLShaderMaterial.empty
-        ),
-        lightPasses = arrayOf(
-            MGMLightPass(
-                GLShaderLightPass.Builder()
-                    .attachAll()
-                    .build(),
-                "shaders/lightPass/vert.glsl",
-                "shaders/opaque/defer/frag_defer_light_dir.glsl",
-            ),
-            MGMLightPass(
-                GLShaderLightPass.Builder()
-                    .attachColorSpec()
-                    .build(),
-                "shaders/lightPass/vert.glsl",
-                "shaders/lightPass/frag_defer.glsl"
-            ),
-            MGMLightPass(
-                GLShaderLightPass.Builder()
-                    .attachDepth()
-                    .build(),
-                "shaders/lightPass/vert.glsl",
-                "shaders/lightPass/frag_defer_depth.glsl"
-            ),
-            MGMLightPass(
-                GLShaderLightPass.Builder()
-                    .attachNormal()
-                    .build(),
-                "shaders/lightPass/vert.glsl",
-                "shaders/lightPass/frag_defer_normal.glsl"
-            )
-        ),
-        GLShaderLightPassPointLight.Builder()
-            .attachAll()
-            .build()
-    )
-
-    private val mVerticesSphere = GLArrayVertexConfigurator(
-        GLEnumArrayVertexConfiguration.SHORT
-    )
-
-    private val mVerticesBox10 = GLArrayVertexConfigurator(
-        GLEnumArrayVertexConfiguration.BYTE
-    )
-
-    private val mVerticesQuad = GLArrayVertexConfigurator(
-        GLEnumArrayVertexConfiguration.BYTE
-    )
-
-    private val verticesBox10Raw = ASUtilsBuffer.createFloat(
-        MGUtilsVertIndices.createCubeVertices2(
-            0.5f
-        )
-    )
-
-    private val mDrawerBox10 = GLDrawerVertexArray(
-        mVerticesBox10
-    )
-
-    private val mBufferUniformCamera = GLBufferUniformCamera(
-        GLBuffer(
-            GL_UNIFORM_BUFFER
-        )
-    )
-
-    private val mCameraFree = COMatrixTranslate().run {
-        COMCamera(
-            GLCameraFree(
-                COCameraFree(
-                    this
-                ),
-                mHandlerGl,
-                mBufferUniformCamera
-            ),
-            GLCameraProjection(
-                COCameraProjection(
-                    this
-                ),
-                mHandlerGl,
-                mBufferUniformCamera
-            )
-        )
-    }
-
-    private val managerFrustrum = COManagerFrustrum(
-        mCameraFree.projection,
-        COMArrayVertexManager(
-            verticesBox10Raw
-        )
-    )
-
-    private val mFramebufferG = GLFrameBufferG(
-        GLFramebuffer()
-    )
-
-    private val mDrawerVolumes = GLDrawerVolumes(
-        mDrawerBox10,
-        managerFrustrum
-    )
-
-    private val mDrawerLightDirectional = GLDrawerLightDirectional()
-
-    private val mDrawerSphere = GLDrawerVertexArray(
-        mVerticesSphere
-    )
-
-    private val mGeometry = MGMGeometry(
-        ConcurrentLinkedQueue(),
-        ConcurrentLinkedQueue(),
-        MGSky(),
-        GLDrawerVertexArray(
-            mVerticesQuad
-        ),
-        mDrawerSphere
-    )
-
-    private val mParameters = MGMParameters(
-        true,
-        null
-    )
-
-    private val mVolume = MGMVolume(
-        mDrawerVolumes,
-        mParameters
-    )
-
-    private val mPools = MGPoolTextures(
-        MGLoaderTextureAsync(
-            mHandlerGl
-        )
-    ).run {
-        MGMPools(
-            MGPoolMaterials(
-                this,
-                mInformatorShader
-            ),
-            MGPoolMeshesStatic(
-                mHandlerGl
-            ),
-            this
-        )
-    }
-
-    private val managers = MGMManagers(
-        LGManagerProcessTime(),
-        GLDrawerLights(
-            GLDrawerLightPass(
-                arrayOf(
-                    mFramebufferG.textureAttachmentPosition.texture,
-                    mFramebufferG.textureAttachmentNormal.texture,
-                    mFramebufferG.textureAttachmentColorSpec.texture,
-                    mFramebufferG.textureAttachmentMisc.texture,
-                    mFramebufferG.textureAttachmentDepth.texture,
-                ),
-                mDrawerSphere
-            )
-        ),
-        managerFrustrum,
-        LGManagerTriggerMesh()
-    )
-
-
-    private val mSwitcherDrawMode = MGSwitcherDrawMode(
-        mFramebufferG,
-        mInformatorShader,
-        mGeometry,
-        mVolume,
-        mDrawerLightDirectional,
-        GLDrawerFramebufferG(
-            mFramebufferG
-        ),
-        managers.managerLight
-    )
-
-    private val mHud = APHud(
-        mCameraFree.camera,
-        requesterUserContent,
-        mSwitcherDrawMode,
-        mParameters,
-        mPools,
-        mInformatorShader,
-        managers,
-        mGeometry,
-        mHandlerGl
-    )
-
-    val sensors = arrayListOf(
-        MGSensorGyroscope(
-            mCameraFree.camera
-        )
-    )
-
     init {
-        val scriptLightPlacement = SCScriptLightPlacement(
-            COUtilsFile.getPublicFile(
-                "scripts"
-            ),
-            managers.managerProcessTime,
-            managers.managerLight,
-            managers.managerFrustrum
-        )
-        scriptLightPlacement.execute()
-
-        mHandlerGl.post(
-            object: COIRunnableBounds {
-                override fun run(
-                    width: Int,
-                    height: Int
-                ) {
-                    mFramebufferG.generate(
-                        width, height
-                    )
-                    mHud.layout(
-                        width.toFloat(),
-                        height.toFloat()
-                    )
-                }
-            }
-        )
-
         mHandlerGl.registerCycleTask(
             MGRunnableCycleSwitcherDrawMode(
                 mSwitcherDrawMode
@@ -335,142 +95,11 @@ class APRendererHandler(
     private var mWidth = 0
     private var mHeight = 0
 
-    fun stop() {
-        managers.managerProcessTime.apply {
-            stop()
-            unregisterAll()
-        }
-    }
-
     override fun onSurfaceCreated(
         gl: GL10?,
         config: EGLConfig?
     ) {
         MGUtilsFile.glWriteExtensions()
-
-        mBufferUniformCamera.apply {
-            buffer.generate()
-            GLBufferUniform.setupBindingPoint(
-                0,
-                buffer,
-                sizeBytes
-            )
-        }
-
-        mVerticesQuad.configure(
-            ASUtilsBuffer.createFloat(
-                MGUtilsVertIndices.createQuadVertices()
-            ),
-            ASUtilsBuffer.createByte(
-                MGUtilsVertIndices.createQuadIndices()
-            ),
-            GLPointerAttribute.Builder()
-                .pointPosition2()
-                .pointTextureCoordinates()
-                .build()
-        )
-
-        mInformatorShader.wireframe.setup(
-            mBuffer,
-            APFile(
-                "shaders/opaque/vert.glsl"
-            ),
-            APFile(
-                "shaders/wireframe/frag_defer.glsl"
-            ),
-            GLBinderAttribute.Builder()
-                .bindPosition()
-                .build()
-        )
-
-        GLBinderAttribute.Builder()
-            .bindPosition()
-            .bindTextureCoordinates()
-            .build().apply {
-                mInformatorShader.lightPasses.forEach {
-                    it.shader.setup(
-                        mBuffer,
-                        APFile(
-                            it.vertPath
-                        ),
-                        APFile(
-                            it.fragPath
-                        ),
-                        this
-                    )
-                }
-
-                mInformatorShader.lightPassPointLight.setup(
-                    mBuffer,
-                    APFile(
-                        "shaders/lightPass/vert_pointLight.glsl"
-                    ),
-                    APFile(
-                        "shaders/opaque/defer/frag_defer_light_point.glsl"
-                    ),
-                    this
-                )
-            }
-
-        mGeometry.meshSky.configure(
-            mInformatorShader,
-            mPools.textures
-        )
-
-        val pointPosition = GLPointerAttribute.Builder()
-            .pointPosition()
-            .build()
-
-        ASObject3d.createFromAssets(
-            "objs/sphere.fbx"
-        )?.get(0)?.let {
-            mVerticesSphere.configure(
-                it.vertices,
-                it.indices,
-                GLPointerAttribute.Builder()
-                    .pointPosition()
-                    .pointTextureCoordinates()
-                    .pointNormal()
-                    .build()
-            )
-        }
-
-        val indicesBox = ASUtilsBuffer.createByte(
-            MGUtilsVertIndices.createCubeIndices2()
-        )
-
-
-        mVerticesBox10.configure(
-            verticesBox10Raw,
-            indicesBox,
-            pointPosition
-        )
-
-        managers.managerProcessTime.apply {
-            registerLoopProcessTime(
-                managerFrustrum
-            )
-            start()
-        }
-
-        mCameraFree.projection.run {
-            modelMatrix.setPosition(
-                0f, 0f, 0f
-            )
-            modelMatrix.invalidatePosition()
-        }
-
-        SCLoaderScripts.executeDirLight(
-            mDrawerLightDirectional
-        )
-
-        glDepthFunc(
-            GL_LESS
-        )
-
-        glCullFace(
-            GL_BACK
-        )
     }
 
     override fun onSurfaceChanged(
@@ -479,11 +108,6 @@ class APRendererHandler(
         height: Int
     ) {
         Log.d(TAG, "onSurfaceChanged: ${Thread.currentThread().name}")
-
-        mCameraFree.projection.setPerspective(
-            width / 2,
-            height
-        )
 
         mWidth = width
         mHeight = height
@@ -500,10 +124,4 @@ class APRendererHandler(
             mWidth, mHeight
         )
     }
-
-    fun onTouchEvent(
-        event: MotionEvent
-    ) = mHud.touchEvent(
-        event
-    )
 }
